LPS-70725 Remove duplicated logic - select field type
